rentconnect/requirements.txt
asgiref==3.8.1
distlib==0.3.9
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
mysql-connector-python==9.3.0
platformdirs==4.3.6
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.1
virtualenv==20.29.3
virtualenvwrapper-win==1.2.7
